agent/requirements.txt
langgraph>=1.0.0
langchain-groq>=1.0.0
langchain-core>=1.0.0
python-dotenv>=1.0.1
